Moved the collision2D class to the Utilities folder.
